refactor: change codeId type from i32 to u64 in contract.create and update related entry function

+++ examples/test/assembly/runtime.ts
@@ -12,7 +12,7 @@ namespace contract {
   ): void;
 
   export declare function create(
-    codeId: i32,
+    codeId: u64,
     initArgs: ArrayBuffer
   ): ArrayBuffer;
 }
